reorganized goal conditioned tasks
